requirements.txt
attrs==16.3.0
Automat==0.5.0
cffi==1.9.1
constantly==15.1.0
futures==3.0.5
incremental==16.10.1
Logbook==1.0.0
lxml==3.7.3
py==1.4.32
pyasn1==0.1.7
pycparser==2.17
pycrypto==2.6.1
pygame==1.9.3
pymodbus==1.2.0
pymunk==4.0.0
pyserial==2.7
pyshark==0.3.6.3
pytest==3.0.7
PyYAML==3.12
scadasim==0.1
six==1.10.0
trollius==1.0.4
Twisted==15.0.0
zope.interface==4.1.2
